Added Parser and other math stuffs

+++ src/RowEquivalent.h
@@ -5,8 +5,8 @@ private:
     MatrixSolver(){}
     int *rowMap;
     int numOfEq;
-    float *Matrix;
-    float *solutionRow;
+    double *Matrix;
+    double *solutionRow;
     void printMat(){
         for(int i=0;i<numOfEq;i++){
             for(int j=0;j<numOfEq+1;j++)
@@ -21,21 +21,21 @@ private:
     }
 
    void normalize(int row){
-        float temp = Matrix[rowMap[row]*(numOfEq+1)+row];
+        double temp = Matrix[rowMap[row]*(numOfEq+1)+row];
         for(int j=row;j<=numOfEq;j++){
             Matrix[rowMap[row]*(numOfEq+1)+j] = Matrix[rowMap[row]*(numOfEq+1)+j]/temp;
         }
     }
     void spreadChangeBelow(int row){
         for(int j=row+1;j<numOfEq;j++){
-            float temp = Matrix[rowMap[j]*(numOfEq+1)+row];
+            double temp = Matrix[rowMap[j]*(numOfEq+1)+row];
             for(int k=row;k<=numOfEq;k++)
                 Matrix[rowMap[j]*(numOfEq+1)+k] = Matrix[rowMap[j]*(numOfEq+1)+k] - temp*Matrix[rowMap[row]*(numOfEq+1)+k];
         }
     }
     void spreadChangeAbove(int row){
         for(int j=0;j<=row-1;j++){
-            float temp = Matrix[rowMap[j]*(numOfEq+1)+row];
+            double temp = Matrix[rowMap[j]*(numOfEq+1)+row];
             for(int k=row;k<=numOfEq;k++)
                 Matrix[rowMap[j]*(numOfEq+1)+k] = Matrix[rowMap[j]*(numOfEq+1)+k] - temp*Matrix[rowMap[row]*(numOfEq+1)+k];
 
@@ -43,7 +43,7 @@ private:
     }
 
 public:
-    bool setMatrix(float* matrix,int numOfEquations, float *solnRow){
+    bool setMatrix(double* matrix,int numOfEquations, double *solnRow){
         Matrix = matrix;
         numOfEq = numOfEquations;
         rowMap = new int[numOfEquations];
